Move all remaining private functions from RigWellTargetMapping to RigWellTargetMappingTools
Moves findStartCell, growCluster, findCandidates, getActiveCellCount,
loadVectorByName, loadOilVectorByName, loadGasVectorByName,
generateStatistics, computeStatisticsAndCreateVectors,
accumulateResultsForSingleCase, and computeBoundingBoxForResult.

RigWellTargetMapping now has no private section; all worker logic
lives in RigWellTargetMappingTools. All 39 unit tests pass.
