+++ ApplicationLibCode/ReservoirDataModel/Well/RigWellTargetMapping.h
@@ -18,21 +18,13 @@
 
 #pragma once
 
-#include "cafVecIjk.h"
-
-#include "cvfBoundingBox.h"
 #include "cvfStructGrid.h"
 
 #include "RigEclipseResultAddress.h"
 
-#include <list>
-#include <map>
-#include <optional>
-#include <utility>
+#include <vector>
 
-class RigActiveCellInfo;
 class RigCaseCellResultsData;
-class RigMainGrid;
 class RimEclipseCase;
 class RimEclipseCaseEnsemble;
 class RimRegularGridCase;
@@ -158,69 +150,5 @@ public:
         double permeability;
         double pressure;
     };
-
-private:
-    static std::optional<caf::VecIjk0> findStartCell( RimEclipseCase*            eclipseCase,
-                                                      size_t                     timeStepIdx,
-                                                      const VolumeType           volumeType,
-                                                      const ClusteringLimits&    limits,
-                                                      const DataContainer&       data,
-                                                      const std::vector<double>& filterVector,
-                                                      const std::vector<int>&    clusters );
-
-    static void growCluster( RimEclipseCase*            eclipseCase,
-                             const caf::VecIjk0&        startCell,
-                             const VolumeType           volumeType,
-                             const ClusteringLimits&    limits,
-                             const DataContainer&       data,
-                             const std::vector<double>& filterVector,
-                             std::vector<int>&          clusters,
-                             int                        clusterId,
-                             size_t                     timeStepIdx,
-                             int                        maxIterations );
-
-    static std::vector<size_t> findCandidates( RimEclipseCase*            eclipseCase,
-                                               const std::vector<size_t>& previousCells,
-                                               const VolumeType           volumeType,
-                                               const ClusteringLimits&    limits,
-                                               const DataContainer&       data,
-                                               const std::vector<double>& filterVector,
-                                               std::vector<int>&          clusters );
-
-    static std::optional<size_t> getActiveCellCount( RimEclipseCase* eclipseCase );
-
-    static std::vector<RigWellTargetMapping::ClusterStatistics> generateStatistics( RimEclipseCase*            eclipseCase,
-                                                                                    const std::vector<double>& pressure,
-                                                                                    const std::vector<double>& permeabilityX,
-                                                                                    int                        numClustersFound,
-                                                                                    size_t                     timeStepIdx,
-                                                                                    const QString&             clusterResultName );
-
-    static void computeStatisticsAndCreateVectors( RimEclipseCase&                         targetCase,
-                                                   const QString&                          resultName,
-                                                   const std::vector<std::vector<double>>& vec );
-
-    static void accumulateResultsForSingleCase( RimEclipseCase&                                      eclipseCase,
-                                                RimEclipseCase&                                      targetCase,
-                                                std::map<QString, std::vector<std::vector<double>>>& resultNamesAndSamples,
-                                                std::vector<int>&                                    occupancy,
-                                                size_t                                               timeStepIdx );
-
-    static cvf::BoundingBox computeBoundingBoxForResult( RimEclipseCase& eclipseCase, const QString& resultName, size_t timeStepIndex );
-
-    static std::vector<double> loadVectorByName( RigCaseCellResultsData& resultsData, const QString& resultName, size_t timeStepIdx );
-
-    static std::vector<double> loadOilVectorByName( RigCaseCellResultsData&    resultsData,
-                                                    VolumesType                volumesType,
-                                                    VolumeResultType           volumeResultType,
-                                                    size_t                     timeStepIdx,
-                                                    const RigFloodingSettings& floodingSettings );
-
-    static std::vector<double> loadGasVectorByName( RigCaseCellResultsData&       resultsData,
-                                                    RiaDefines::EclipseUnitSystem unitsType,
-                                                    VolumesType                   volumesType,
-                                                    VolumeResultType              volumeResultType,
-                                                    size_t                        timeStepIdx,
-                                                    const RigFloodingSettings&    floodingSettings );
 
 };
